requirements.txt
mmcv-full==1.5.0
mmdet==2.23.0
fvcore==0.1.5
matplotlib==3.7.1
numpy==1.23.5
pandas==1.3.4
pycocotools==2.0.6
reliability==0.8.13
scikit_learn==1.0.1
scipy==1.10.0
setuptools==65.6.3
shortuuid==1.0.11
terminaltables==3.1.10
torch==1.11.0
tqdm==4.62.3
